Created handleCardClick() function and reset_game() function in script.js.
handleCardClick() - check if the card clicked is invisible or not, if it's invisible then make it
visible and set the status to pending and check the secong card. if both
the cards are same then change the status to done and check another card.
If not, then make both cards invisible and remove the status.

reset_game() - checks if any of the card is visible then make it invisible
and removes the status of the cards. It also resets the variables.